# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fabware API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{id}/revisions:
    get:
      operationId: listRevisions
      tags: [parts]
      summary: List all design revisions for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Revision history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PartRevision"
  /projects/{id}/revisions/{revisionId}:
    get:
      operationId: getRevision
      tags: [parts]
      summary: Get a single revision with its rendered preview (no DB write)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: revisionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Revision detail with preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevisionPreview"
  /projects/{id}/revisions/{revisionId}/restore:
    post:
      operationId: restoreRevision
      tags: [parts]
      summary: Restore a specific revision as the active part spec
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: revisionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated part
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevisionChangeResponse"
  /projects/{id}/revisions/undo:
    post:
      operationId: undoRevision
      tags: [parts]
      summary: Move to the previous revision
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated part
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevisionChangeResponse"
  /projects/{id}/revisions/redo:
    post:
      operationId: redoRevision
      tags: [parts]
      summary: Move to the next revision
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated part
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevisionChangeResponse"
...
components:
  schemas:
...
    PartSpec:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        partType:
          type: string
          description: "e.g. bracket, plate, enclosure, channel, angle"
        material:
          type: ["string", "null"]
        thickness:
          type: ["number", "null"]
        width:
          type: ["number", "null"]
        height:
          type: ["number", "null"]
        depth:
          type: ["number", "null"]
        bendRadius:
          type: ["number", "null"]
        bendAngles:
          type: ["string", "null"]
          description: JSON array of bend angles in degrees
        holePattern:
          type: ["string", "null"]
          description: JSON describing hole locations and diameters
        powderCoat:
          type: ["boolean", "null"]
        powderCoatColor:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        svgPreview:
          type: ["string", "null"]
          description: SVG string for 2D flat pattern preview
        sendCutSendUrl:
          type: ["string", "null"]
        dslJson:
          type: ["string", "null"]
          description: Canonical parametric DSL (JSON-encoded) describing the part
        featureGraphJson:
          type: ["string", "null"]
          description: Generated feature graph (JSON-encoded) used to drive preview and DXF
        currentRevisionId:
          type: ["integer", "null"]
        totalRevisions:
          type: integer
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - partType
        - totalRevisions
        - updatedAt
...
    PartRevision:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        revisionNumber:
          type: integer
        dslJson:
          type: string
        rationale:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - revisionNumber
        - dslJson
        - createdAt
    RevisionPreview:
      type: object
      properties:
        revision:
          $ref: "#/components/schemas/PartRevision"
        partSpec:
          $ref: "#/components/schemas/PartSpec"
        isCurrent:
          type: boolean
      required:
        - revision
        - partSpec
        - isCurrent
    RevisionChangeResponse:
      type: object
      properties:
        partSpec:
          $ref: "#/components/schemas/PartSpec"
        currentRevisionId:
          type: ["integer", "null"]
        canUndo:
          type: boolean
        canRedo:
          type: boolean
      required:
        - partSpec
        - canUndo
        - canRedo
...